requirements.txt
numpy==1.23.2
pkg_resources==0.0.0
pygame==2.1.2
PyOpenGL==3.1.6
